Fixes after actually trying the script

+++ setup_scripts/install_golang.py
@@ -3,11 +3,13 @@
 # source code is governed by a BSD-style license that can be found in the
 # LICENSE file.
 
+import argparse
 import os
 import shutil
 import subprocess
 import tarfile
 import urllib2
+import sys
 
 
 def from_sources():
@@ -36,22 +38,36 @@ def from_sources():
 def from_binary():
   """Install globally from pre-built binaries as root."""
   goroot = '/usr/local/go'
-  shutil.rmtree(goroot)
+  print('Installing Go in %s' % goroot)
+  if os.path.isdir(goroot):
+    shutil.rmtree(goroot)
   # TODO(maruel): Magically figure out latest version.
-  version = '1.7.3'
+  version = '1.7.4'
   uname = os.uname()[4]
   arch = 'amd64'
   if uname.startswith('arm'):
-    arch = 'arm6l'
+    # ~70MB, at 1Mbit it takes 12 minutes...
+    arch = 'armv6l'
   if sys.platform == 'linux2':
     os_name = 'linux'
   filename = 'go' + version + '.' + os_name + '-' + arch + '.tar.gz'
   # TODO(maruel): It sucks that it doesn't validate TLS certificate.
-  o = urllib2.urlopen('https://storage.googleapis.com/golang/' + filename)
-  with tarfile.open(fileobj=o) as t:
-    # TODO(maruel): It sucks that tarfile doesn't do the right tihng.
-    t.extractall(path=goroot)
-  os.remove(filename)
+  url = 'https://storage.googleapis.com/golang/' + filename
+  print('Fetching %s' % url)
+  with tarfile.open(fileobj=urllib2.urlopen(url), mode='r|gz') as t:
+    i = 0
+    while True:
+      n = t.next()
+      if not n:
+        break
+      if n.name.startswith('.', '/') or '..' in n.name:
+        raise Exception('Dangerous tar file')
+      t.extract(n, goroot)
+      # There's over 5650 files so print a dot at every 100 files.
+      if not (i%100):
+        sys.stdout.write('.')
+        sys.stdout.flush()
+      i += 1
   with open('/etc/profile.d/golang.sh', 'wb') as f:
     f.write('export PATH="$PATH:%s/bin"\n' % goroot)
   os.chmod('/etc/profile.d/golang.sh', 0555)
@@ -59,7 +75,7 @@ def from_binary():
 
 def main():
   parser = argparse.ArgumentParser(description=sys.modules[__name__].__doc__)
-  parser.add_option('--system', action='store_true')
+  parser.add_argument('--system', action='store_true')
   args = parser.parse_args()
   if args.system:
     from_binary()
